Rename Renderer to ResponseSerializer and inline to full_stack_spec
